test: Create a header file for TestUtils
Put this class definition in a header file so it can be referenced
elsewhere.
